Updated the test files in this directory with values that were calculated in an NCL code
